reorganizing gettr bot build / source files to maintain paradigm shared by other bots
