+++ src/bot/gettr/gettr.hpp
